Checkpoint commit - Lexical analysis
The base structure for the lexer/tokenizer classes are now defined. check the previous commit.

there's a need for:
identifiers string and list of strings read for keywords like: allow_methods, etc...

NOTE: maybe look up to a better way of handling the Lexer::get_identifier() method using a hash-map.